LICENCE-MIT: Add the MIT text
Generated with:

  $ curl -s https://raw.githubusercontent.com/rust-lang/rust/1.23.0/LICENSE-MIT >LICENSE-MIT

Rust dropped their copyright line from LICENSE-MIT in
rust-lang/rust@2a8807e8 (LICENSE-MIT: Remove inaccurate
(misattributed) copyright notice, 2017-07-26, [1]).  I'm not clear on
how that interacts with the "The above copyright notice and this
permission notice shall be included..." line [2], but it's better than
not including the MIT text at all (which is where we were before this
commit).  Josh Triplett responded to my concerns with:

On Thu, Jan 11, 2018 at 08:04:39PM +0000, Josh Triplett wrote [3]:
> I don't believe so, no. The requirement is to reproduce what's
> there; if something isn't there you don't need to reproduce it.

which is good enough for me.

[4] has the same text (although it has a copyright template), but
using Rust as the source gets us the Rust-standard line wrappings ;).

[1]: https://github.com/rust-lang/rust/pull/43498/commits/2a8807e889a43c6b89eb6f2736907afa87ae592f
[2]: https://github.com/rust-lang/rust/pull/43498#discussion_r161025446
[3]: https://github.com/rust-lang/rust/pull/43498#discussion_r161061389
[4]: https://opensource.org/licenses/MIT